Revert "hp bar"
This reverts commit e8ff48e23a84ad536849938faa3a9e7e7b804b17.

+++ SpaceCOG Client/Assets/Scripts/enemyScript.cs
@@ -68,7 +68,7 @@ public class enemyScript : MonoBehaviour {
 	}
 
 	public void Damage(float damage) {
-		hp -= (int) damage;
+		hp -= damage;
 		Debug.Log ("Enemy lost" + damage + "hp.  Enemy now at " + hp + "hp");
 		if (hp <= 0) {
 			Debug.Log ("Enemy destroyed");
